test(hooks): colocate value control hook tests
